requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
openai>=0.27.0
anthropic>=0.5.0
google-generativeai>=0.3.0